Split CollectionMarshaller into bounded vs unbounded implementations.
