+++ clustering/marshalling/protostream/src/main/java/org/wildfly/clustering/marshalling/protostream/util/BoundedCollectionMarshaller.java
@@ -24,17 +24,18 @@ package org.wildfly.clustering.marshalling.protostream.util;
 
 import java.util.Collection;
 import java.util.Map;
-import java.util.function.Function;
+import java.util.function.IntFunction;
 
-import org.wildfly.clustering.marshalling.protostream.ProtoStreamMarshaller;
+import org.wildfly.clustering.marshalling.protostream.PrimitiveMarshaller;
+import org.wildfly.clustering.marshalling.spi.ConstantFunction;
 
 /**
- * Collection marshaller for collections constructed with a context.
+ * Collection marshaller for bounded collections.
  * @author Paul Ferraro
  */
-public class ContextualCollectionMarshaller<T extends Collection<Object>, C> extends CollectionMarshaller<T, C, C> {
+public class BoundedCollectionMarshaller<T extends Collection<Object>> extends CollectionMarshaller<T, Void, Integer> {
 
-    public ContextualCollectionMarshaller(Class<T> targetClass, Function<C, T> factory, Function<T, C> context, ProtoStreamMarshaller<C> contextMarshaller) {
-        super(targetClass, factory, Map.Entry::getKey, context, contextMarshaller);
+    public BoundedCollectionMarshaller(Class<T> targetClass, IntFunction<T> factory) {
+        super(targetClass, factory::apply, Map.Entry::getValue, new ConstantFunction<>(null), PrimitiveMarshaller.VOID.cast(Void.class));
     }
 }
