+++ clustering/marshalling/protostream/src/main/java/org/wildfly/clustering/marshalling/protostream/util/CollectionMarshaller.java
@@ -22,20 +22,78 @@
 
 package org.wildfly.clustering.marshalling.protostream.util;
 
+import java.io.IOException;
+import java.util.AbstractMap;
 import java.util.Collection;
 import java.util.Map;
-import java.util.function.IntFunction;
+import java.util.OptionalInt;
+import java.util.function.Function;
 
-import org.wildfly.clustering.marshalling.protostream.PrimitiveMarshaller;
-import org.wildfly.clustering.marshalling.spi.ConstantFunction;
+import org.infinispan.protostream.ImmutableSerializationContext;
+import org.infinispan.protostream.RawProtoStreamReader;
+import org.infinispan.protostream.RawProtoStreamWriter;
+import org.wildfly.clustering.marshalling.protostream.ObjectMarshaller;
+import org.wildfly.clustering.marshalling.protostream.Predictable;
+import org.wildfly.clustering.marshalling.protostream.ProtoStreamMarshaller;
 
 /**
- * Collection marshaller for collections constructed with a capacity.
+ * ProtoStream optimized collection marshaller.
  * @author Paul Ferraro
  */
-public class CollectionMarshaller<T extends Collection<Object>> extends AbstractCollectionMarshaller<T, Void, Integer> {
+public class CollectionMarshaller<T extends Collection<Object>, C, CC> implements ProtoStreamMarshaller<T> {
 
-    public CollectionMarshaller(Class<T> targetClass, IntFunction<T> factory) {
-        super(targetClass, factory::apply, Map.Entry::getValue, new ConstantFunction<>(null), PrimitiveMarshaller.VOID.cast(Void.class));
+    private final Class<T> targetClass;
+    private final Function<CC, T> factory;
+    private final Function<Map.Entry<C, Integer>, CC> constructorContext;
+    private final Function<T, C> context;
+    private final ProtoStreamMarshaller<C> contextMarshaller;
+
+    public CollectionMarshaller(Class<T> targetClass, Function<CC, T> factory, Function<Map.Entry<C, Integer>, CC> constructorContext, Function<T, C> context, ProtoStreamMarshaller<C> contextMarshaller) {
+        this.targetClass = targetClass;
+        this.factory = factory;
+        this.constructorContext = constructorContext;
+        this.context = context;
+        this.contextMarshaller = contextMarshaller;
+    }
+
+    @Override
+    public T readFrom(ImmutableSerializationContext context, RawProtoStreamReader reader) throws IOException {
+        C collectionContext = this.contextMarshaller.readFrom(context, reader);
+        int size = reader.readUInt32();
+        CC constructorContext = this.constructorContext.apply(new AbstractMap.SimpleImmutableEntry<>(collectionContext, size));
+        T collection = this.factory.apply(constructorContext);
+        for (int i = 0; i < size; ++i) {
+            collection.add(ObjectMarshaller.INSTANCE.readFrom(context, reader));
+        }
+        return collection;
+    }
+
+    @Override
+    public void writeTo(ImmutableSerializationContext context, RawProtoStreamWriter writer, T collection) throws IOException {
+        C collectionContext = this.context.apply(collection);
+        this.contextMarshaller.writeTo(context, writer, collectionContext);
+        writer.writeUInt32NoTag(collection.size());
+        for (Object element : collection) {
+            ObjectMarshaller.INSTANCE.writeTo(context, writer, element);
+        }
+    }
+
+    @Override
+    public OptionalInt size(ImmutableSerializationContext context, T collection) {
+        C collectionContext = this.context.apply(collection);
+        OptionalInt size = this.contextMarshaller.size(context, collectionContext);
+        if (size.isPresent()) {
+            size = OptionalInt.of(size.getAsInt() + Predictable.unsignedIntSize(collection.size()));
+            for (Object element : collection) {
+                OptionalInt elementSize = ObjectMarshaller.INSTANCE.size(context, element);
+                size = size.isPresent() && elementSize.isPresent() ? OptionalInt.of(size.getAsInt() + elementSize.getAsInt()) : OptionalInt.empty();
+            }
+        }
+        return size;
+    }
+
+    @Override
+    public Class<? extends T> getJavaClass() {
+        return this.targetClass;
     }
 }
